CAMEL-13108: Move mock component out of camel-core
