+++ camel-core/src/main/java/org/apache/camel/processor/SendProcessor.java
@@ -28,8 +28,8 @@ import org.apache.camel.ExchangePattern;
 import org.apache.camel.RuntimeCamelException;
 import org.apache.camel.Traceable;
 import org.apache.camel.impl.DefaultProducerCache;
-import org.apache.camel.impl.InterceptSendToEndpoint;
 import org.apache.camel.spi.IdAware;
+import org.apache.camel.spi.InterceptSendToEndpoint;
 import org.apache.camel.spi.ProducerCache;
 import org.apache.camel.support.AsyncProcessorSupport;
 import org.apache.camel.support.EndpointHelper;
